+++ src/main/java/com/example/soopgwan/domain/habit/domain/Day.java
@@ -1,0 +1,36 @@
+package com.example.soopgwan.domain.habit.domain;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.time.LocalDate;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
+@Table(name = "tbl_day")
+@Entity
+public class Day {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(columnDefinition = "DATE", nullable = false)
+    private LocalDate startAt;
+
+    @Column(columnDefinition = "DATE", nullable = false)
+    private LocalDate endAt;
+
+    @Column(columnDefinition = "VARCHAR(200)", nullable = false)
+    private String eval;
+}
